requirements.txt
numpy<2.0.0
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.3
structlog==23.2.0
matplotlib==3.7.2
pandas==2.0.3
mistralai==0.0.11
